requirements.txt
openant==1.2.1
pygame==2.1.3
pyusb==1.2.1

